+++ src/OmniSharp.DotNetTest/Models/RunTestsInContextRequest.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+using OmniSharp.Mef;
+
+namespace OmniSharp.DotNetTest.Models
+{
+    [OmniSharpEndpoint(OmniSharpEndpoints.V2.RunTestsInContext, typeof(RunTestsInContextRequest), typeof(RunTestResponse))]
+    public class RunTestsInContextRequest : BaseTestsInContextRequest
+    {
+    }
+}