+++ src/OmniSharp.MSBuild/ProjectFile/ProjectFileInfo.cs
@@ -41,6 +41,8 @@ namespace OmniSharp.MSBuild.ProjectFile
         public string DocumentationFile => _data.DocumentationFile;
         public ImmutableArray<string> PreprocessorSymbolNames => _data.PreprocessorSymbolNames;
         public ImmutableArray<string> SuppressedDiagnosticIds => _data.SuppressedDiagnosticIds;
+        public ImmutableArray<string> WarningsAsErrors => _data.WarningsAsErrors;
+        public ImmutableArray<string> WarningsNotAsErrors => _data.WarningsNotAsErrors;
 
         public bool SignAssembly => _data.SignAssembly;
         public string AssemblyOriginatorKeyFile => _data.AssemblyOriginatorKeyFile;
@@ -52,9 +54,12 @@ namespace OmniSharp.MSBuild.ProjectFile
         public ImmutableArray<PackageReference> PackageReferences => _data.PackageReferences;
         public ImmutableArray<string> Analyzers => _data.Analyzers;
         public ImmutableArray<string> AdditionalFiles => _data.AdditionalFiles;
+        public ImmutableArray<string> AnalyzerConfigFiles => _data.AnalyzerConfigFiles;
         public ImmutableDictionary<string, string> ReferenceAliases => _data.ReferenceAliases;
         public ImmutableDictionary<string, string> ProjectReferenceAliases => _data.ProjectReferenceAliases;
         public bool TreatWarningsAsErrors => _data.TreatWarningsAsErrors;
+        public bool RunAnalyzers => _data.RunAnalyzers;
+        public bool RunAnalyzersDuringLiveAnalysis => _data.RunAnalyzersDuringLiveAnalysis;
         public string DefaultNamespace => _data.DefaultNamespace;
 
         public ProjectIdInfo ProjectIdInfo { get; }
@@ -75,7 +80,7 @@ namespace OmniSharp.MSBuild.ProjectFile
         {
             var id = ProjectId.CreateNewId(debugName: filePath);
 
-            return new ProjectFileInfo(new ProjectIdInfo(id, isDefinedInSolution:false), filePath, data: null);
+            return new ProjectFileInfo(new ProjectIdInfo(id, isDefinedInSolution: false), filePath, data: null);
         }
 
         internal static ProjectFileInfo CreateNoBuild(string filePath, ProjectLoader loader)
@@ -83,7 +88,7 @@ namespace OmniSharp.MSBuild.ProjectFile
             var id = ProjectId.CreateNewId(debugName: filePath);
             var project = loader.EvaluateProjectFile(filePath);
             var data = ProjectData.Create(project);
-            //we are not reading the solution here 
+            //we are not reading the solution here
             var projectIdInfo = new ProjectIdInfo(id, isDefinedInSolution: false);
 
             return new ProjectFileInfo(projectIdInfo, filePath, data);
@@ -125,7 +130,7 @@ namespace OmniSharp.MSBuild.ProjectFile
 
             var data = ProjectData.Create(projectInstance);
             var projectFileInfo = new ProjectFileInfo(ProjectIdInfo, FilePath, data);
-            var eventArgs = new ProjectLoadedEventArgs(Id, projectInstance, diagnostics, isReload: true, ProjectIdInfo.IsDefinedInSolution,data.References);
+            var eventArgs = new ProjectLoadedEventArgs(Id, projectInstance, diagnostics, isReload: true, ProjectIdInfo.IsDefinedInSolution, data.References);
 
             return (projectFileInfo, diagnostics, eventArgs);
         }