+++ src/OmniSharp.MSBuild/ProjectManager.cs
@@ -26,6 +26,7 @@ using OmniSharp.Services;
 using OmniSharp.Utilities;
 using System.Reflection;
 using Microsoft.CodeAnalysis.Diagnostics;
+using OmniSharp.Roslyn.EditorConfig;
 
 namespace OmniSharp.MSBuild
 {
@@ -176,7 +177,7 @@ namespace OmniSharp.MSBuild
                     await Task.Delay(LoopDelay, cancellationToken);
                     ProcessQueue(cancellationToken);
                 }
-                catch(Exception ex)
+                catch (Exception ex)
                 {
                     _logger.LogError($"Error occurred while processing project updates: {ex}");
                 }
@@ -377,6 +378,30 @@ namespace OmniSharp.MSBuild
                 QueueProjectUpdate(projectFileInfo.FilePath, allowAutoRestore: true, projectFileInfo.ProjectIdInfo);
             });
 
+            if (_workspace.EditorConfigEnabled)
+            {
+                // Watch beneath the Project folder for changes to .editorconfig files.
+                _fileSystemWatcher.Watch(".editorconfig", (file, changeType) =>
+                {
+                    QueueProjectUpdate(projectFileInfo.FilePath, allowAutoRestore: false, projectFileInfo.ProjectIdInfo);
+                });
+
+                // Watch in folders above the Project folder for changes to .editorconfig files.
+                var parentPath = Path.GetDirectoryName(projectFileInfo.FilePath);
+                while (parentPath != Path.GetPathRoot(parentPath))
+                {
+                    if (!EditorConfigFinder.TryGetDirectoryPath(parentPath, out parentPath))
+                    {
+                        break;
+                    }
+
+                    _fileSystemWatcher.Watch(Path.Combine(parentPath, ".editorconfig"), (file, changeType) =>
+                    {
+                        QueueProjectUpdate(projectFileInfo.FilePath, allowAutoRestore: false, projectFileInfo.ProjectIdInfo);
+                    });
+                }
+            }
+
             if (projectFileInfo.RuleSet?.FilePath != null)
             {
                 _fileSystemWatcher.Watch(projectFileInfo.RuleSet.FilePath, (file, changeType) =>
@@ -436,14 +461,21 @@ namespace OmniSharp.MSBuild
             UpdateReferences(project, projectFileInfo.ProjectReferences, projectFileInfo.References);
             UpdateAnalyzerReferences(project, projectFileInfo);
             UpdateAdditionalFiles(project, projectFileInfo.AdditionalFiles);
+            UpdateAnalyzerConfigFiles(project, projectFileInfo.AnalyzerConfigFiles);
             UpdateProjectProperties(project, projectFileInfo);
 
             _workspace.TryPromoteMiscellaneousDocumentsToProject(project);
-            _workspace.UpdateDiagnosticOptionsForProject(project.Id, projectFileInfo.GetDiagnosticOptions());
+            _workspace.UpdateCompilationOptionsForProject(project.Id, projectFileInfo.CreateCompilationOptions());
         }
 
         private void UpdateAnalyzerReferences(Project project, ProjectFileInfo projectFileInfo)
         {
+            if (!projectFileInfo.RunAnalyzers || !projectFileInfo.RunAnalyzersDuringLiveAnalysis)
+            {
+                _workspace.SetAnalyzerReferences(project.Id, ImmutableArray<AnalyzerFileReference>.Empty);
+                return;
+            }
+
             var analyzerFileReferences = projectFileInfo.Analyzers
                 .Select(analyzerReferencePath => new AnalyzerFileReference(analyzerReferencePath, _analyzerAssemblyLoader))
                 .ToImmutableArray();
@@ -477,10 +509,32 @@ namespace OmniSharp.MSBuild
 
             foreach (var file in additionalFiles)
             {
-                 if (File.Exists(file))
-                 {
+                if (File.Exists(file))
+                {
                     _workspace.AddAdditionalDocument(project.Id, file);
-                 }
+                }
+            }
+        }
+
+        private void UpdateAnalyzerConfigFiles(Project project, IList<string> analyzerConfigFiles)
+        {
+            if (!_workspace.EditorConfigEnabled)
+            {
+                return;
+            }
+
+            var currentAnalyzerConfigDocuments = project.AnalyzerConfigDocuments;
+            foreach (var document in currentAnalyzerConfigDocuments)
+            {
+                _workspace.RemoveAnalyzerConfigDocument(document.Id);
+            }
+
+            foreach (var file in analyzerConfigFiles)
+            {
+                if (File.Exists(file))
+                {
+                    _workspace.AddAnalyzerConfigDocument(project.Id, file);
+                }
             }
         }