+++ test-assets/test-projects/NUnitTestProject/TestProgram.cs
@@ -9,7 +9,7 @@ namespace Main.Test
         public void Test()
         {
             Assert.True(true);
-        }
+        } // This is here for boundary testing
 
         [TestCase(0)]
         [TestCase(1)]
@@ -42,7 +42,7 @@ namespace Main.Test
         {
             int a = 1, b = 1;
             Console.WriteLine($"a = {a}, b = {b}");
-            Assert.AreEqual(a,b);
+            Assert.AreEqual(a, b);
         }
 
         public void UtilityFunction()
@@ -51,5 +51,39 @@ namespace Main.Test
         }
 
         private static int[] _items = new int[1] { 1 };
+    }
+
+    [TestFixture(typeof(int))]
+    [TestFixture(typeof(double))]
+    public class GenericTest<T>
+    {
+        [Test]
+        public void TypedTest()
+        {
+            Assert.NotNull(default(T));
+        }
+
+        [Test]
+        public void TypedTest2()
+        {
+            Assert.Null(default(T));
+        }
+
+        private class NestedClass
+        {
+
+            public void M()
+            {
+
+            }
+        }
+    }
+
+    public class NoTests
+    {
+        public void M()
+        {
+
+        }
     }
 }